Migrating Video to shoes/types
* Converting `Video` to move into `shoes/types`.
* REMINDER: need to remove `shoes/video/video.c` from `file_list`, old
SOLOCS are now in `env.rb`. Merge with care.
* Shoes3/shoes3#333
